requirements.txt
flask
pymongo
python-dotenv
flask-cors
resend
cryptography
dnspython
email-validator
pytest
